% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{additional_parameters}
\alias{additional_parameters}
\title{Additional parameters for indicators}
\format{
A data frame with indicator additional parameters in rows and columns:
\itemize{
  \item{\code{indicator_name}: String with indicator name. Should not contain spaces (use underscore for multiple word separation).}
  \item{\code{parameter}: String with parameter name, as defined in the indicator function.}
  \item{\code{default_value}: Default value for the parameter.}
  \item{\code{description}: String describing the function parameter.}
}
}
\description{
A data frame with the description of additional parameters for indicator functions
}
\examples{
data(additional_parameters)
}
\seealso{
\code{\link{indicator_definition}}, \code{\link{variable_definition}}
}
\keyword{data}
