% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_indicators.R
\name{available_indicators}
\alias{available_indicators}
\title{Available indicator information}
\usage{
available_indicators()
}
\value{
A data frame with information of available indicator functions and their requirements.
}
\description{
Provides the list of indicators currently available
}
