% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_indicators.R
\name{estimate_indicators}
\alias{estimate_indicators}
\title{Estimate forest indicators}
\usage{
estimate_indicators(
  indicators,
  stand_static_input,
  stand_dynamic_input = NA,
  plant_static_input = NA,
  plant_dynamic_input = NA,
  timber_volume_function = NA,
  plant_biomass_function = NA,
  additional_params = list(),
  verbose = TRUE
)
}
\arguments{
\item{indicators}{A character vector containing the indicators to be estimated.}

\item{stand_static_input}{A data frame (or \code{\link{sf}} object) containing static stand variables. Minimum required column are \code{id_stand} (character).}

\item{stand_dynamic_input}{Optional data frame containing dynamic stand variables. Minimum required columns are \code{id_stand} (character) and \code{date} (\code{\link{Date}}).}

\item{plant_static_input}{Optional data frame containing static plant variables. Minimum required columns are \code{id_stand} (character) and \code{plant_entity} (character).}

\item{plant_dynamic_input}{Optional data frame containing dynamic plant variables. Minimum required columns are \code{id_stand} (character), \code{plant_entity} (character) and \code{date} (\code{\link{Date}}).}

\item{timber_volume_function}{Optional function supplied for timber volume calculation.}

\item{plant_biomass_function}{Optional function supplied for whole-plant biomass calculation.}

\item{additional_params}{Optional named list where each element is in turn a list of additional parameters required for internal indicator functions.}

\item{verbose}{A logical flag to provide information on progress}
}
\value{
A data frame with the following columns:
  \itemize{
    \item{\code{id_stand} - Stand identifier.}
    \item{\code{date} - Date of indicator assessment.}
    \item{\code{...} - Additional columns with names equal to strings given in \code{indicators}.}
  }
}
\description{
Estimates forest indicators for a given set of forest stands during a given set of dates of evaluation.
}
\details{
Additional details...
}
