% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestables2forestindicators.R
\name{forestables2forestindicators}
\alias{forestables2forestindicators}
\title{Wrapper function for structures obtained with forestables}
\usage{
forestables2forestindicators(
  x,
  country = "ES",
  version = "ifn2",
  type = "plant_dynamic_input"
)
}
\arguments{
\item{x}{A data frame obtained with package forestables}

\item{country}{Country of the forest inventory (i.e. "ES", "FR" or "US")}

\item{version}{Spanish National Forest Inventory version (i.e. "ifn2", "ifn3" or "ifn4")}

\item{type}{A string of the input data frame required for forestindicators.}
}
\value{
A data frame with data structure suitable for forestindicators
}
\description{
Wrapper function for structures obtained with forestables
}
