% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestindicators-package.R
\docType{package}
\name{forestindicators-package}
\alias{forestindicators-package}
\alias{forestindicators}
\title{forestindicators: estimation of multiple indicators of forest ecosystem services}
\description{
Facilitates the estimation multiple indicators of forest ecosystem services. Indicators are defined in internal functions and tables.
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/forestindicators/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
[ORCID](https://orcid.org/0000-0001-7132-2080)

Author: Núria Aquilué
[ORCID](https://orcid.org/0000-0001-7911-3144)

Contributor: Víctor Granda
[ORCID](https://orcid.org/0000-0002-0469-1991)

Contributor: José Salgado

Contributor: Giuseppe Cappizzi
}
\keyword{internal}
