% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formes2forestindicators.R
\name{formes2forestindicators}
\alias{formes2forestindicators}
\title{Wrapper function for simulations with package FORMES}
\usage{
formes2forestindicators(x, initial_year = NA, type = "plant_dynamic_input")
}
\arguments{
\item{x}{An object issued from simulations with functions \code{\link[FORMES]{IFNscenario}}, 
\code{\link[DSSsim]{run_formes_scenario}}.}

\item{initial_year}{A numeric value of the initial year (step 0 of the simulation)}

\item{type}{A string describing the input data frame for forestindicators.}
}
\value{
A data frame with data structure suitable for forestindicators
}
\description{
Wrapper function for simulations with package FORMES
}
