% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{indicator_definition}
\alias{indicator_definition}
\title{Indicator definition and requirements}
\format{
A data frame with parameters in rows and columns:
\itemize{
  \item{\code{indicator_name}: String with indicator name. Should not contain spaces (use underscore for multiple word separation).}
  \item{\code{stand_static_variables}: .}
  \item{\code{stand_dynamic_input}: .}
  \item{\code{stand_dynamic_variables}: .}
  \item{\code{plant_static_input}: .}
  \item{\code{plant_static_variables}: .}
  \item{\code{plant_dynamic_input}: .}
  \item{\code{plant_dynamic_variables}: .}
  \item{\code{output_units}: .}
  \item{\code{estimation_method}: .}
  \item{\code{interpretation}: .}
  \item{\code{implementation_responsible}: .}
}
}
\description{
A data frame of implemented indicator metrics and their requirements
}
\examples{
data(indicator_definition)
}
\seealso{
variable_definition
}
\keyword{data}
