% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfate2forestindicators.R
\name{medfate2forestindicators}
\alias{medfate2forestindicators}
\title{Wrapper function for simulations with package medfate or medfateland}
\usage{
medfate2forestindicators(x, id_stand = NA, type = "plant_dynamic_input")
}
\arguments{
\item{x}{An object issued from simulations with functions \code{\link[medfate]{fordyn}}, \code{\link[medfateland]{fordyn_spatial}},
\code{\link[medfateland]{fordyn_scenario}} or \code{\link[medfateland]{fordyn_land}}}

\item{id_stand}{A string identifying the forest stand (if \code{x} has been obtained using \code{\link[medfate]{fordyn}}).}

\item{type}{A string describing the input data frame for forestindicators.}
}
\value{
A data frame with data structure suitable for forestindicators
}
\description{
Wrapper function for simulations with package medfate or medfateland
}
