% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_information.R
\name{show_information}
\alias{show_information}
\title{Describe an individual indicator}
\usage{
show_information(indicator)
}
\arguments{
\item{indicator}{Character string of the indicator to be described.}
}
\value{
The function returns console output only.
}
\description{
Prompts information in to the console regarding the definition, units and data requirements for a given
forest indicator.
}
\examples{
show_information("timber_harvest")
}
\seealso{
\code{\link{available_indicators}}
}
