% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{variable_definition}
\alias{variable_definition}
\title{Input variable definitions}
\format{
A data frame with parameters in rows and columns:
\itemize{
  \item{\code{variable_name}: String with variable name. Should not contain spaces (use underscore for multiple word separation).}
  \item{\code{level}: .}
  \item{\code{units}: .}
  \item{\code{description}: .}
}
}
\description{
A data frame of variable definitions and units, as used in indicator estimation
}
\examples{
data(variable_definition)
}
\seealso{
indicator_definition
}
\keyword{data}
