% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{example_datasets}
\alias{example_datasets}
\alias{example_stand_static_input}
\alias{example_stand_dynamic_input}
\alias{example_plant_static_input}
\alias{example_plant_dynamic_input}
\title{Example datasets for indicator estimation}
\description{
Datasets provided as examples of data input to illustrate indicator estimation and for
indicator function testing.
}
\examples{
data(example_stand_static_input)
}
\seealso{
indicator_definition
}
\keyword{data}
