% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{variable_definition}
\alias{variable_definition}
\title{Input variable definitions}
\format{
A data frame with variables in rows and columns:
\itemize{
  \item{\code{variable_name}: String with variable name. Should not contain spaces (use underscore for multiple word separation).}
  \item{\code{level}: String describing the level (scale) to which the variable applies, either 'stand' or 'plant'.}
  \item{\code{type}: String with variable type, either 'character' or 'numeric'.}
  \item{\code{units}: String indicating the variable units.}
  \item{\code{description}: String with the verbal description of the variable.}
}
}
\description{
A data frame of variable definitions and units, as used in indicator estimation
}
\examples{
data(variable_definition)
}
\seealso{
\code{\link{indicator_definition}}, \code{\link{additional_parameters}}
}
\keyword{data}
